Display time range in session details
