+++ src/app/conf/2025/schedule/_components/format-block-time.tsx
@@ -1,0 +1,13 @@
+import { parseISO } from "date-fns"
+
+const timeFormat = new Intl.DateTimeFormat(undefined, {
+  hour: "2-digit",
+  minute: "2-digit",
+})
+export const formatBlockTime = (start: string, end?: Date) => {
+  const startDate = parseISO(start)
+  if (end) {
+    return timeFormat.formatRange(startDate, end)
+  }
+  return timeFormat.format(startDate)
+}
