+++ src/app/conf/2025/schedule/_components/schedule-list.tsx
@@ -16,6 +16,7 @@ import {
   FilterStates,
   ResetFiltersButton,
 } from "./filters"
+import { formatBlockTime } from "./format-block-time"
 
 export interface FiltersConfig
   extends Partial<
@@ -99,19 +100,6 @@ function getSessionsByDay(
       }),
     ),
   }
-}
-
-const timeFormat = new Intl.DateTimeFormat(undefined, {
-  hour: "2-digit",
-  minute: "2-digit",
-})
-const formatBlockTime = (start: string, end?: string) => {
-  const startDate = parseISO(start)
-  if (end) {
-    const endDate = parseISO(end)
-    return timeFormat.formatRange(startDate, endDate)
-  }
-  return timeFormat.format(startDate)
 }
 
 export interface ScheduleListProps {
@@ -205,8 +193,18 @@ export function ScheduleList({
                 </h3>
                 {Object.entries(concurrentSessionsGroup).map(
                   ([sessionDate, sessions], i, blocks) => {
-                    const blockEnd = sessions[0]?.event_end
-                    const nextBlockStart = blocks[i + 1]?.[0]
+                    const blockEnd = new Date(
+                      Math.max(
+                        ...sessions.map(session =>
+                          new Date(session.event_end).getTime(),
+                        ),
+                      ),
+                    )
+
+                    const nextBlock = blocks[i + 1]
+                    const nextBlockStart = nextBlock?.[0]
+                      ? new Date(nextBlock[0])
+                      : undefined
 
                     const isBreak =
                       sessions[0]?.event_type
